Initial Commit
This project has three subcomponents:
1. A scheduler (located under the minis folder)
2. A simulator (Simulator.py)
3. Graphical Monitors (in graphics/)

Scheduler:
A Visit is a single exposure. A VisitPair is two visits to the same
ra/dec. A MiniSurvey is a collection of VisitPairs that tiles the sky
with a particular overlapping pattern. An OngoingSurvey is a collection
of MiniSurveys. The scheduler has a working set of VisitPairs calculated
in _calculatePendingVisitPairs drawn from the VisitPairs currently in
the OngoingSurvey. It then divides these visits into scans that go from
dec=0 to the south pole, and it schedules visits in each scan in such a
way that revisits can happen without having to slew far.

Simulator:
The simulator runs the scheduler and currently calculates some
rudimentary statistics about the generated schedule. It also keeps track
of the current time.

Graphical Monitor:
The main graphical interface is in GraphicalMonitor. The Simulator keeps
a GraphicalMonitor around, and it notifies it of each visit that the
Scheduler schedules. The GraphicalMonitor then updates a display to show
the sky being covered by the chosen schedule.

Issues:
1. There's also a performance issue where old MiniSurveys are
kept around when they have only a few visits left in them, and
_calculatePendingVisitPairs has to loop over all remaining MiniSurveys.
I think what I should do is make the MiniSurvey class entirely static
(so never instantiate it) and then not group visits into MiniSurveys.
That way, the Scheduler can keep VisitPairs in a set and then just
remove completed VisitPairs entirely.

2. The scheduler does not obey the telescope's maximum altitude
constraint. It also ignores the fact that every time the schedule passes
by zenith, the dome has to rotate 180 degrees. My planned solution is to
look only to the North of zenith on some nights and only to the South of
zenith on other nights.

3. Slew times are being underestimated. I need to figure out exactly how
opsim calculates slew times and replicate that.