Scheduler can schedule only in the North or South
Scheduler divides the sky into Telescope.latitude \pm 5 degrees and
schedules observations for the night only in the North or only in the
South (the survey region now extends farther to the North than the
equator in Config.py). Problem is we get a hole for decs close to
Telescope.latitude. Will try to address this by making the North/South
cutoff in azimuth instead of in declination.