Bit of a graphics overhall
Sorry for the large commit.
1. Now batch process addVisit in the _addVisits function. This doesn't
actually batch process anything in batch yet (just loops through the
pending visits that were added with addVisit), but it might be useful
in the future (and I was planning to batch process but it didn't pan out).
2. Updated the limiting x/y values from wcs_world2pix to be more
accurate so it no longer has padding around the sky.
3. Changed wcs coordinate transform calls to have origin 1 but still not
sure if this is right (but it sends (0,0) to (0,0) so that's good?)
4. Keep track of where pixels are in skyPix using rowLengthsCumSum so
that we can efficiently find candidate pix that are covered by a visit
(before I was using a KDTree on the pixels in 3D cartesian coords which
was too slow).
5. Deleted old, commented out lsst_sims code that was much slower than
the current implementation.