Look East
The scheduler now looks "East" (i.e. between decs of
Telescope.latitude-zenithBuffer and Telescope.latitude+zenithBuffer).

The telescope can't keep up with the sky if we include -90<dec<30 in
the WFD survey (which is ~31,000 degrees) and if it tries to get all
decs every night. So to cut down on slew time, it chooses either North
or SouthEast every night and looks only in that direction. If North, it
scans up and down on the meridian to the north of the zenith avoidance
zone. If SouthEast, it alternates between scanning the meridian to the
South and catching fields that lie in the zenith buffer zone (in the
East) before they hit zenith.