Added loneliness and percentile maps
SkyMap can now compute a loneliness map, which is a map of the fraction
of visits to a pixel unaccompanied by another visit within a specified
number of minutes.

It can also compute a percentile map, which is a map where the color of
each pixel represents the nth percentile of time between two visits to
that pixel.