Scheduler can now generate scans in two ways
Way one: each scan has a fixed RA range. A disadvantage is this way
seems to cause us to spend more time slewing all the way to the South
pole and back. An advantage is that it's simple.

Way two: slice the celestial sphere with parallel planes parallel to the
meridian at the middle of the night (I think this is how I did it? It
was a while ago). Each scan consists of the visits that fall between
each consecutive pair of planes. This seems like it would make us slew
to the pole less, but it fails for nights close to 12 hours long.