Quick fix: scrap rest of night if ever cloudy
Scheduler doesn't know that if it's cloudy for 6 hours it should skip
enough scans so it's again observing on the meridian. Without this fix,
it starts observing below the horizon if it's been cloudy for long
enough.

The real fix is to have the Scheduler know to skip scans if it notices
the dome has been closed for a while.