Always schedule visits from tonight's new mini
Replaced `ongoingSurvey` with `makeupVPs`. Each night, if there are
enough visits in makeupVPs, the night is devoted to executing only those
visit pairs. On all other nights, a new minisurvey is scheduled along
with any remaining visits taken from makeupVPs. At the end of the night
the visits not executed from the new mini are added to makupVPs.

The advantage of this is that on non-makeup nights, fewer unnecessary
revisits are performed to the same pixels. I was also hoping it would
reduce slew times but I'm not sure if that happened.

A disadvantage seems to be that the makeup nights become more extreme,
causing the scheduler to schedule too many visits in the same scans on a
given night (which means it strays very far from the meridian).