Bugfix: Scheduler was ending night unnecessarily
The Simulator is now responsible for ending the night (by calling
Scheduler.notifyNightEnd()), so the Scheduler shouldn't return because
it thinks the night is over (or that there's no more observing for the
night b/c of clouds). Instead, it should yield None until it has more
visits to schedule.