Add rudimentary deep drilling support
At most one DD observation is carried out per night. The DD field chosen
on any given night is just the first DD field listed in ddFields.csv
that falls in the correct RA/dec regions for the night. If there is no
such DD field in range, then no DD observation is carried out that
night.

Part of the refactoring needed for DD support was to introduce the
revisitGroup abstraction, which is a list of two scans with directions
on how to carry them out twice (visit and revisit). On each night, the
revisitGroup whose RA range contains the night's DD visit is identified.
The DD visit is executed immediately after this revisitGroup is
completed.

With this implementation, DD visits are given about 4% of the survey
time.