Refactored some scheduling code to NightScheduler
A NightScheduler is a scheduler that can schedule a single night. It
takes in the night number that it's scheduling, the direction of the
night, the telescope being used, and a set of makeup visit pairs that it
can draw from. At the end of the night, it returns (via notifyNightEnd)
a set of visit pairs that should be added to makeup VPs.

I think this refactoring is an improvement from the previous situation
where everything was crammed into Scheduler but it has several problems:
1. There is too much coupling between Scheduler and NightScheduler.
NightScheduler.__init__ takes four parameters, and the shuffling of
VPs between Scheduler and NightScheduler via makeupVPs is very shady.
2. It would be ideal if the NightScheduler could have two subclasses:
NorthNightScheduler and SoutheastNightScheduler, since right now there
are just two methods (one for each direction) that are called based on
which direction the night is. However, there's enough shared
functionality between North and Southeast to make this awkward.
3. RevisitGroups should still be refactored out.