Refactored MiniSurveys into Tilings
MiniSurvey is now responsible for rotating the tilings and cutting them
in RA/Dec.

Tiling.py is responsible for generating some fixed tiling (e.g. 2-raft
overlap tiling or opsim-like tiling) given a desired density.
I added a Thomson tiling with approximate solutions to the Thomson
problem for N up to 10000 so that MiniSurvey can choose N to get
about the right density (which is set in config.py).